Authorizer: Avoid cancelling the log parser on success
Instead, let the log parser stop itself when it sees MARKER_CODE_STOP,
which will always be present for the success flow.

The previous behavior of explicitly cancelling the log parser when
Rcbridge.rbAuthorize() returns is unfortunately prone to a race
condition. It turns out that logging from the golang side is not
synchronous, so the log message containing the token would sometimes
occur after the cancellation message.

Fixes: #55

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>