rcbridge: Avoid calling Android functions over JNI
Since go 1.22.5, this results in a panic. There is an upstream bug
report at [1] for musl, but the same thing happens with bionic. The same
fix of using pthread_getattr_np() on bionic also works, but it's better
to avoid needing folks to recompile the go toolchain.

We were just using these calls as a small hack to pass a list of structs
to the Java side, which we can easily do another way.

[1] https://github.com/golang/go/issues/68285

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>