Make dynamic shortcuts configurable
This way, the user can choose which remotes they want shortcuts for.
Previously, the visibility was tied to whether external app access was
blocked. The default state for this new option is to not add a shortcut.
The state is saved in rclone.conf so it is preserved when exporting and
reimporting a config.

This commit also fixes RSAF crashing when the number of shortcuts
exceeds Android's limit (usually 15). If there are too many shortcuts,
the list will now be silently truncated. There is no UI warning for this
because the user-facing limit is the launcher's limit (usually 4), which
is way less than Android's limit and there's no API for querying it. It
doesn't make sense to show a warning long after the launcher already
started hiding shortcuts.

Fixes: #65

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>