Add support for disabling VFS caching
VFS caching can now be configured per-remote. Disabling it is equivalent
to `rclone mount --vfs-cache-mode none`. This will disable support for
random writes and prevent failed uploads from being retried. However, it
allows uploads to begin immediately and allows the client to show more
accurate file write progress.

If a remote type does not support streaming, VFS caching will be force
enabled and the UI option will be grayed out. This is preferable to
allowing the user to set a broken configuration and not knowing until
they try to create a file and having it fail.

Internally, RSAF now maintains two sets of VFS instances. One set with
caching VFSs and another set with streaming VFSs. This allows the
caching option to be toggled immediately without waiting for open files
to be closed (and fully uploaded). The downside is that if the user
toggles the option while uploads are occurring, the directory listings
may be inconsistent until the uploads complete.

Issue: #79

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>