rcbridge: Increase polling interval for VFS cache cleaner
The background upload monitor service only keeps the process alive long
enough for all files to be closed. However, for the uploaded files to be
deleted from the cache directory, we need to wait one additional polling
cycle. Instead of keeping the service alive longer, let's just reduce
the polling interval so that it has a greater chance of completing
before Android SIGSTOPs us.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>