RcloneProvider: Block rename, copy, and move while source file is open
Android's StorageManager.openProxyFileDescriptor() makes it impossible
to have a blocking close() call from the client's point of view. This
causes problems with client apps that write data to a temporary file and
then rename it when complete. Rclone's VFS cache writeback process fails
if a file gets renamed while it is being uploaded.

Since the asynchronicity of close() is something we're stuck with, the
best we can do is make other operations blocking. RcloneProvider now
tracks which files are open for writing and forces the rename, copy, and
move operations to wait if the source file is one of the open files or
any of its parent directories. This is sufficient to make common file
access patterns work, but if a client app tries to shoot itself in the
foot, it can absolutely still do so.

Fixes: #81

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>