Add support for hiding files on a remote when the app is locked
This commit adds a new per-remote option for hiding files while the app
is locked. This is implemented as an extension of the existing allow
external access toggle. However, with this option:

* The SAF roots remain visible
* File operations throw FileNotFoundException instead of
  SecurityException

The idea is that external apps are temporarily blocked from access, but
still retain all previously granted permissions.

Unlike the original proposed design, queryChildDocuments() will also
throw FileNotFoundException instead of returning an empty list. This is
so the same function is used for enforce all security restrictions and
make it easy to verify the correctness of the logic.

When the inactivity period expires, the restrictions are enforced for
new file operations only. Ongoing operations, like file copies/moves,
and reads/writes against already-open file descriptors are still
allowed.

Fixes: #106

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>