Add support for DocumentsProvider.openDocumentThumbnail()
This will return a pipe and then spawn a background task to generate a
thumbnail and write it to the pipe. Cancellation is supported in between
logical operations (eg. parse headers, load frame, etc.). Android has no
way to interrupt the parsers in MediaMetadataRetriever and ImageDecoder
though, so there will be a delay before cancellation takes effect. This
can be noticeable when scrolling quickly, for example through a list of
videos, since there is a hard concurrency limit equal to the number of
logical cores.

Note that only audio, image, and video formats natively supported by
Android can be thumbnailed. We don't bundle any decoders.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>