Add support for resuming VFS cache writeback
Previously, if the VFS cache writeback was interrupted, it would
effectively not be resumable, especially if the files were big. After
the interruption, the first file operation on remotes with pending
uploads would block until the uploads completed. However, there would be
no foreground service to keep the process alive. The only way to get out
of the situation would be to leave the RSAF settings screen open until
the uploads invisibly complete.

With this commit, BackgroundUploadMonitorService is now also started on
boot or when the user first opens the app after it crashes or is killed.
It will scan through the VFS cache directory and initialize the VFS
instances for remotes with pending uploads. The service now exits on its
own after a period of time with no VFS cache files open.

For this to work well, the rclone VFS is configured so that writebacks
are asynchronous. This is necessary to prevent VFS initialization from
blocking until uploads are complete, but unfortunately also makes the
close() operation asynchronous. RSAF is no longer able to show a
notification if a background upload fails.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>