Remove support for POSIX-like semantics
This turned out to be not such a great idea. Applications expect
document providers to behave like Android's builtin FileSystemProvider
for local files. POSIX-like semantics differed significantly from that
and there are no known applications that explicitly try to make use of
this functionality. Let's get rid of this feature and simplify the code.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>