Add proper support for Android's CA trust stores
golang does not support Android's various trust stores well. The CA
certificates in /system/etc/security/cacerts are the only ones that are
properly read. The updatable apex trust store in Android 14+ and the
user trust store are ignored. System CA certificates that the user
disables are still permitted to be used.

This commit implements our own certificate loading mechanism that loads
from all of Android's trust stores and respects disabled certificates.
We generate a combined PEM file in a temp file and feed that to rclone's
CaCert option.

While live reloading is supported, the user experience is not that
great. Due to rclone's caching, when certificates are reloaded, they
only take effect in new remotes. The user will need to delete and
recreate the remote, export and reimport, or force close RSAF.

Fixes: #119

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>