Prompt before actions that delete the VFS cache
Previously, when deleting a remote, renaming (copy + delete) a remote,
or importing a configuration, we'd wait up to 1 minute for pending VFS
writeback operations to complete and then delete the remote and VFS
cache. This was a terrible user experience because:

* the user does not have the choice to wait for file operations to
  complete if they weren't already aware of them
* the UI would appear as if nothing was happening
* the user would have to wait for no reason if they don't care about
  pending uploads

With this commit, we'll show a prompt before destructive operations that
delete the VFS cache. If the user chooses to proceed, the deletion
occurs immediately. This is not perfect though--a VFS writeback thread
cannot be cancelled and will likely throw a checksum validation error
when it completes.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>