Only allow copy/move via rclone when it can be done server-side
If copyDocument() and moveDocument() are slow, the binder transaction
will time out and Android will kill RSAF. If we can't perform the
operation quickly via server-side operations, let's just tell the client
that the operation is unsupported. Android's DocumentsUI specifically
handles this scenario and will do the copy/move operation itself, which
isn't subject to this problem.

Fixes: #124

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>