rcbridge: Fix absolute path appearing in .go.buildinfo ELF section
Even though we build with -trimpath, the absolute path to the source
directory still appears in the .go.buildinfo ELF section. This is
because gomobile generates its own go.mod file with a `replace`
directive that points to an absolute path. The path comes from parsing
the output of `go list -m -json all`. There is no way to change this
behavior in `gomobile bind`.

To work around this, this commit introduces a very hacky wrapper around
the go executable. When running `go list -json ...`, the absolute path
in the output will be replaced with a relative path starting at the base
path specified by GOWRAPPER_BASE_PATH. `gomobile bind` always runs the
affected command from its temporary build directory:

    ${TMPDIR}/gomobile-work-<RANDOM>/src-android-${ARCH}

Since we already control TMPDIR and set it to a fixed location, we can
set GOWRAPPER_BASE_PATH to this full path and the resulting relative
path embedded into the .go.buildinfo section will be the following,
regardless of where it is built:

    ../../../../../../../rcbridge

This removes one more source of non-reproducibility.

Issue: #50

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>