RcloneProvider: Work around Android StorageManagerService watchdog restart
The watchdog calls StorageManagerService.monitor(), which requires
holding the main vold lock that is also held by the FUSE unmount
operation. The unmount operation is blocked until our FUSE release
callback completes, which could take an indefinite amount of time since
it's waiting for file uploads. The watchdog ends up killing
system_server, which appears to the user as if the device had rebooted.

To work around this, we'll just run the actual rclone close() operation
in a thread. Client apps never see a blocking close() call anyway, so
there is no observable difference.

Fixes: #157

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>